Add ConfigurationVariablesLoader and tests
