+++ install-dev/controllers/http/database.php
@@ -24,6 +24,7 @@
  * @license   https://opensource.org/licenses/OSL-3.0 Open Software License (OSL 3.0)
  */
 
+use PrestaShop\PrestaShop\Core\Util\ConfigurationVariablesLoader;
 use PrestaShopBundle\Install\Database;
 use Symfony\Component\Yaml\Yaml;
 
@@ -108,7 +109,6 @@ class InstallControllerHttpDatabase extends InstallControllerHttp implements Htt
         $this->session->database_password = trim(Tools::getValue('dbPassword'));
         $this->session->database_prefix = trim(Tools::getValue('db_prefix'));
         $this->session->database_clear = Tools::getValue('database_clear');
-        $this->session->use_env_variables = Tools::getValue('use_env_variables');
         $this->session->rewrite_engine = Tools::getValue('rewrite_engine');
     }
 
@@ -199,6 +199,9 @@ class InstallControllerHttpDatabase extends InstallControllerHttp implements Htt
                 $parameters = Yaml::parse(file_get_contents(_PS_ROOT_DIR_ . '/app/config/parameters.yml.dist'));
             }
 
+            $configurationVariablesLoader = new ConfigurationVariablesLoader();
+            $parameters = $configurationVariablesLoader->loadEnvVariables($parameters);
+
             $this->database_server = $parameters['parameters']['database_host'];
             if (!empty($parameters['parameters']['database_port'])) {
                 $this->database_server .= ':' . $parameters['parameters']['database_port'];
@@ -210,7 +213,6 @@ class InstallControllerHttpDatabase extends InstallControllerHttp implements Htt
             $this->database_prefix = $parameters['parameters']['database_prefix'];
 
             $this->database_clear = true;
-            $this->use_env_variables = true;
             $this->use_smtp = false;
             $this->smtp_encryption = 'off';
             $this->smtp_port = 25;
@@ -222,7 +224,6 @@ class InstallControllerHttpDatabase extends InstallControllerHttp implements Htt
             $this->database_engine = $this->session->database_engine;
             $this->database_prefix = $this->session->database_prefix;
             $this->database_clear = $this->session->database_clear;
-            $this->use_env_variables = $this->session->use_env_variables;
 
             $this->use_smtp = $this->session->use_smtp;
             $this->smtp_encryption = $this->session->smtp_encryption;
