+++ tests/Unit/Core/Util/ConfigurationVariablesLoaderTest.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace Tests\Unit\Core\Util;
+
+use PHPUnit\Framework\TestCase;
+
+class ConfigurationVariablesLoaderTest extends TestCase
+{
+
+    public function testLoadEnvVariables() :void{
+
+    }
+
+}
